requirements.txt
phidata
openai
psycopg[binary]
pgvector
python-dotenv
duckduckgo-search
pypdf
nest_asyncio
